[refactor] Move some projects out of sandbox
